fix: Cleaner and More Reliable State Machine
The finite state machine which tracks the login state was
refactored. The code is cleaner now. And the view have
fewer glitches. Previously, the view got stuck when the
application was logged in. The state transition was not
handled correctly. A reason was that the state handling
was spread across several classes with change signalling.
This complex code was prone to timing errors.

Signed-off-by: Philipp Le <philipp@philipple.de>