fix: Browser View for Login
The OpenID Connect login is handled in a web view (for mobile platforms)
or external browser (other platforms).

On Android, it is necessary to allow loading unencrypted HTTP pages.
This is fixed with adaptations of the Android Manifest.

Furthermore, previously on Android platforms, the login page was loaded
in an external browser. Now, the In-App View is used.

Signed-off-by: Philipp Le <philipp@philipple.de>